# implement cache abstraction and unit test