--- a/nodedb.go
+++ b/nodedb.go
@@ -29,6 +29,7 @@ const (
 	// Using semantic versioning: https://semver.org/
 	defaultStorageVersionValue = "1.0.0"
 	fastStorageVersionValue    = "1.1.0"
+	fastNodeCacheLimit = 100000
 )
 
 var (
@@ -96,7 +97,7 @@ func newNodeDB(db dbm.DB, cacheSize int, opts *Options) *nodeDB {
 		opts:           *opts,
 		latestVersion:  0, // initially invalid
 		nodeCache:      cache.New(cacheSize),
-		fastNodeCache:  cache.New(100000),
+		fastNodeCache:  cache.New(fastNodeCacheLimit),
 		versionReaders: make(map[int64]uint32, 8),
 		storageVersion: string(storeVersion),
 	}