@p0mvn 2022-03-29
Yes, that's expected for now. I'm planning on fixing the linter as soon as I have some extra time